設定回傳數值已成功為匯入全部資料 缺少d3

+++ database/migrations/2019_04_22_023102_create_smart_watering_table.php
@@ -18,7 +18,7 @@ class CreateSmartWateringTable extends Migration
             $table->string('formname');
             $table->string('sensorname');
             $table->string('mode');
-            $table->integer('vaule');
+            $table->double('vaule');
             $table->dateTime('sendtime');
         });
     }
